requirements.txt
boto3>=1.34.0
sagemaker>=2.200.0
transformers>=4.35.0
torch>=2.1.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
datasets>=2.14.0
python-dotenv>=1.0.0